# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ATLAS Backend API",
        "description": "API documentation for ATLAS Backend endpoints.",
        "version": "1.0.0"
    },
    "paths": {
        "/api/auth/permissions": {
            "get": {
                "tags": [
                    "Permissions"
                ],
                "summary": "Get all permissions",
                "description": "Retrieve a list of all permissions with their associated roles",
                "operationId": "c6f4be15fe8e14a061028344d803e768",
                "responses": {
                    "200": {
                        "description": "List of permissions retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "permissions": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "edit-posts"
                                                    },
                                                    "guard_name": {
                                                        "type": "string",
                                                        "example": "web"
                                                    },
                                                    "roles_list": {
                                                        "type": "string",
                                                        "example": "Admin, Editor"
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "format": "date-time"
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "format": "date-time"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Ocurrió un error al obtener los permisos."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
        "/api/auth/permissions/dropdown": {
            "get": {
                "tags": [
                    "Permissions"
                ],
                "summary": "Get permissions for dropdown",
                "description": "Retrieve all permissions formatted for dropdown/select components",
                "operationId": "bb92b9118e416dd6c11f9b88dd49c000",
                "responses": {
                    "200": {
                        "description": "List of permissions for dropdown retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "value": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "label": {
                                                "type": "string",
                                                "example": "edit-posts"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Ocurrió un error al obtener los permisos."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Authentication using Laravel Sanctum bearer token",
                "bearerFormat": "Bearer",
                "scheme": "bearer"
            }
        }
    },
...
}